requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-socketio==5.13.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.13
python-dotenv==1.1.1
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary
httpx>=0.27.0
alembic>=1.13.0

